Fix bug #493 (AMReplyLong in vapi-conduit/SEGMENT_EVERYTHING)
No longer use RDMA for ReplyLongM in SEGMENT_{LARGE,EVERYTHING} configs,
since doing so may require an AM round-trip to pin the target.

So, for the FAST configuration the value of gasnet_AMMaxLongReply() remains
the same as gasnet_AMMaxLongRequest() at 2GB-1.  However, for the LARGE
and EVERYTHING configurations the value is slightly less than 4k.