Update the collective notes to refect a new day's set of "deep thoughts"
First I ran it through the spell checker to clean it up.

In this version I've added my proposed interfaces which manage to address
both the local/global knowledge issue and the 1,N,M,N*M,Arbitrary address
count issue with just five "flavors" of each op.  It looks like just using
flags to express these variations was not going to be satisfactory.

So the relocalization collectives will require
 2 (blocking vs. nb) * 5 (the "flavors") = 10 functions per operation