Initial checkin of VIS support
Includes a complete reference implementation of the proposed VIS spec.

The current VIS reference implementation is implemented entirely in terms of puts/gets, and in terms of each other (indexed over vector, strided over indexed and strided over vector). This algorithm is rather slow in the general case (ie it's often what the user was trying to avoid by calling VIS functions in the first place), but it is a correct implementation, and there are some cases where it's the right thing to do (e.g. fully contiguous transfers or a few large chunks on a conduit with fast RDMA but slow AM), so it's a useful algorithm to have available. AM-based reference implementations will come next.

The algorithm selection mechanism is designed for extensibility, allowing conduits to easily override the heuristic and choose an algorithm based on the transfer parameters and current system state, selecting either one of the prepackaged reference implementations, or using conduit-provided native implementations. Currently the default algorithm selection mechanism is pseudo-random (for purposes of testing coverage), but we'll eventually replace this with something smarter once we have AM-based implementations.

Includes a brutal, pseudo-randomized (deterministic) tester that hits most of the VIS usage cases (and has even exposed some minor non VIS-related general GASNet bugs). The test is currently passing on all conduits, except gm-conduit and vapi-conduit with seg-everything, where it's apparently exposed a bug in firehose.

Also..
Add a build-time flag to libgasnet
Add some additional functionality to test.h