partial fix to bug 517 uncovers some problems with GM 2.0 firehose remote callback
The bugfix covers two issues.  For one, dst/src fields passed to the firehose
get function were reversed (!), and this never showed up so far because all our
other tests are symmetric.  I'm currently forcing GM 2.0 to issue a get
operation after a firehose remote request with a callback is completed (we
typically rely on the remote callback to complete the get by doing a "reverse
put").  I plan to look into this further soon, but until then gets that miss in
the firehose table will require an extra one-sided get operation and not make
any use of the remote callback feature.