Reapply all my perlstart changes that Paul so rudely clobberred
And make the one-line typo fix that was the only change required
In my defense, I DID test my changes on four different platforms before committing - I just didn't happen to get a configuration that was missing readlink. I've re-tested this change on four new platforms, including all those reported to be problematic (pcp-i-1 and big mac).

I realize my checkin had a problem and apologize for the inconvenience. However, in the future please give the responsible developer a chance to respond before stomping on such a wide set of changes - 24 hours seems like a reasonable courtesy window to wait for a response to an email-only query, and would have saved me two hours of wasted time undoing all your changes when I could have otherwise fixed the problem in 30 seconds (as it was you only gave me 3 hours of notice and I happened to be away from email). If you absolutely must rollback a non-trivial change without help from the developer, you should always create a cvs tag to allow easy re-applying of the change.

Also remove Paul's changes to the conduit Makefile.am's quadrupling CONDUIT_EXTRALIBCFLAGS variable in favor of the "other" solution (just use the existing CONDUIT_EXTRALIBCFLAGS). GASNET_THREAD_DEFINES intentionally contains just defines, so it's perfectly harmless to apply it twice on the compile line. The point of all these Makefile changes is to keep the set of required conduit Makefile variables as minimal as absolutely possible (to simplify template instantiation and maintenance), so until this causes a fatal problem it's not worth the extra complexity.