Fix (work around) bug #522
It turns out that the GNU variadic macros which cause xlc to complain
are defined but never used when building GASNet.  In fact, it appears
that NOTHING in the file mtl_log.h gets used.

So, the work around is to add '-DH_MTL_LOG_H' to the CFLAGS and thus
use the file's double-include protection to ignore it.  This is kind
of gross and thus only applied when configure finds the
compiler/preprocessor unable to deal with GNU-style variadic macros.

Note that Mellanox has no plans to support xlc as a compiler, so
we're not going to get any help from them on this issue.