Fix bug 210 - upcr configure conduit detection
I've completely overhauled conduit detection in upcr - upcr's configure script now makes an "educated guess" at a superset of what the conduit list will be (based on the --enable flags and knowledge of which conduits are autodetected by GASNet), and the list is later trimmed (at first make invocation) to match those conduits actually detected by the GASNet configure script.

This maintains all the current enable/disable conduit functionality (no change to docs required), eliminates the confusing behaviors that were happening when upcr & gasnet disagreed upon the conduit list, and allows us to move towards a model where eventually we can detect most or all the conduits automatically in GASNet configure (with no help from special variables on systems in a standard configuration), and upcr will automatically adjust to match the GASNet detected list.

Currently the code which "trims" the upcr conduit list issues a warning when one of the autodetected conduits that was not explicitly disabled was not detected (ie, one of mpi, udp or smp failed to autodetect), because these are expected to be available on basically all systems. When we add autodetection of some hardware-specific conduits, we'll probably just remove the warning.