Improved -with-segment-mmap-max option.
Added ability to specify the size as "8GB", "4097MB", etc., instead of
forcing user to provide actual number of bytes.

Added to runtime configure.in, so it shows up in the help.

Note that we should try to get rid of this option (by automatically
handling >2GB sizes w/o user's needing to specify a max size) before we
release.