Temporary fix for bug 172.
This was not a LAPI bug but a bug in the conduit.  I've
done a quick hack to fix the problem until I can get back to
this and upgrade the implementation to use the new LAPI
implementation (before the next release).

On Federation systems, the network packet size if 2KB, on
older switch2-based systems (Seaborg) the packet size is 1KB.
The hack is to detect, at compile time, whether the system is
using the new version of LAPI.  We detect this by the presence of
the LAPI_DGSM_COPY_SIZE macro in the lapi.h file.

This is a hack because all Federation systems are using the new
version of LAPI and the older systems are not.  The code will still
work on an older system even if its using the new LAPI, but space
will be wasted.