Editing pass over the collectives design document, make some design progress.
Incorporate several issues discussed with Paul, and some new issues that he hasn't seen.
Once we finish the design of the computational collectives, the plan is this document can be trimmed down to become the GASNet collectives spec.

Add several notes on the current interface design and the implications for usage and implementation.

Clarify the wording in many places to indicate how given requirements will generalize in the presence of teams.

Correct several constraints referring to "nodes" were the real relevant constraint is over "images".

Add a whole new section summarizing lessons learned from the MPI collectives spec and how they might be used to improve our system.

Paul, please take time to read though all of this and ensure you concur.. comments are appreciated.