Fix a long standing vapi-conduit problem with CLUMPS.
Though I never created a Bugzilla entry for it, vapi-conduit has never been
able to deal with multiple processes per node.  The reason was that in
gasnet_init() we probe for the limit on pinnable memory.  This probe is done
by actually pinning.  Therefore multiple processes would end up splitting the
available memory in an unpredicatable way, though often with the last process
started getting zero pinnable memory.

The endpoint information we already exchange for connections identifies the
network node ID of each GASNet node.  This we can count the number of GASNet
processes on each network node.  We can also distinguish exactly one per
network node (the lowest GASNet node number presently) which will perform the
pinning limit probe and then divide its result by the number of processes on
that node.  For initializing firehose we already perform an all-to-all
communication to share this pinning information.  By doing so earlier than
we previously did, the non-distinguished nodes can read the pinnable memory
limit determined by the distinguished process on the same node.

This solution is good in that it does not add any network communication to
the initialization that was not present before (except that an all-to-all
that was previously only needed for SEGMENT_LARGE and SEGMENT_EVERYTHING is
now also needed for SEGMENT_LARGE).  The one downside is that if the probe
for pinnable memory encountered the mmap size limit or the OS's per-process
memory limit (man 2 getrlimit), rather than physical memory or HCA resource
limits, then we will get a smaller value (after dividing by the number of
process per node) than might actually be pinnable.  This could be resolved
(if needed) with some additional communication to coordinate more probing.
Another approach would be letting all processes on a node probe concurrenty,
and then take the average of their results.  I am still looking at that
possibility.