Initial commit of the experimental spin-poll throttling feature
To use the feature, you must configure with --enable-throttle-poll, otherwise the default behavior is exactly what it has always been. When enabled, it only affects behavior for (GASNET_PAR || GASNETI_CONDUIT_THREADS). I'm still actively working on fixing some known issues with the throttling, so the feature is not yet recommended for general use.

To support this change, I've reorganized our AMPoll functions somewhat, here are the new guidelines:
 gasnet_AMPoll() - public poll function called by the client, throttled and traced
                   should not be called from within GASNet (so we only trace directly user-initiated calls)
 gasneti_AMPoll() - called internally by GASNet, provides throttling (if enabled), but no tracing
 gasnetc_AMPoll() - conduit AM dispatcher, should only be called from gasneti_AMPoll()
Most of the lines in diff for this commit are due to this simple renaming, which after macro expansion makes absolutely no change when the throttle-pollers feature is disabled (other than tracing).

When the feature is enabled, there are two levels of throttling - calls to gasneti_AMPoll() are throttled to ensure that at most one thread is spin-polling at a time. This functionality is already available for all conduits.

The second level of throttling is to ensure that no thread is spin polling when some thread needs to use the network for "useful work" - specifically, for sending messages. This second level requires conduits to wrap any message-sending code which acquires mutexes (shared with the poll path) with calls to the framework macros gasneti_suspend_spinpollers()/gasneti_resume_spinpollers(), which will cause the spin pollers to backoff and give mutex priority to the thread(s) wishing to send.

So far the second-level throttling hooks have been added to:
   udp, mpi, elan
they should be unnecessary in:
   smp and sci - which have no conduit or driver mutexes shared between the AMPoll and send paths
second-level throttling tie-work remains on:
 lapi - I think I know what's required here and am working on it
 vapi and gm - I haven't looked into these yet, need to discuss with Christian & Paul

Issues left to be addressed:
* Effectively provide AM handler concurrency in the presence of spin-poll throttling
* Reduce relocking overhead for the elected spin poller
* Many other miscellaneous improvements soon to come