Fix bug 471 - AMMPI fails assertion on 4 testexit cases.
The assertion failure was due to the fact that gasnet_exit invokes AMMPI_SPMDExit which uses AMMPI control messages to shut down the job, and AMMPI control messages previously were not safe to issue in a handler context (they used the Request mechanism and caused reentrancy into the message dispatcher, leading to a crash).

I've fixed AMMPI control messages to use the reply mechanism,  making them safe in a handler context and fixing the crashes - mpi-conduit now passes all of the testexit cases (except those where the client is in a non-polling spin loop, which requires interrupt-based AM execution to prevent a hang).

I've also added explicit assertions testing for dispatcher reentrancy to make the problem more obvious in the future if it somehow happens again due to other bugs.

Also, fix the GASNet test delay calibration loop to handle systems with very high timer granularity (eg OSF) - this was causing unreported crashes in testlogGP.
Add a timer granularity/overhead output to all the performance-related GASNet tests and warn if using gettimeofday.
Add testcontend to the standard list of par tests