Fix bug 687.
The UTF-8 issues we've encountered can be fixed by using the 'bytes' module.
However, that module isn't in older version of perl (which don't need it,
because they won't support UTF-8 anyway, and thus won't have the bug).

The fix:  we now check at configure time whether 'perl -Mbytes' works (-M forces
a module to be loaded), and if so, GASNET_PERL_BYTESFLAG is defined to
'-Mbytes'.  Our perlstart now allows a PERLSTART_FLAGS shell variable to specify
any command-line flags that should be passed to perl, so for upcc/upcrun we're now
setting

    PERLSTART_FLAGS="@GASNET_PERL_BYTESFLAG@"

before we call PERLSTART.  This is overkill, in the sense that this bug doesn't
show up on all versions of perl, but we can narrow the hack if we need to
later.

So far I haven't seen any trouble using upcrun under UTF-8, so for now I'm not
trying to use the bytes module with it.

Replaced the duplicate work we use for detecting Perl with a central
GASNET_PROG_PERL macro in gasnet/acinclude.m4.  The only thing I had to jettison
was redirecting the output of perl tests to AC_FD_CC--autoconf seemed unhappy to
see this in the acinclude file.  So I'm sending the output to /dev/null instead.