Step 1 of fixing bug #685 - renaming
This commit does not fix the bug or add any additional stability.
If fact, except for Alpha and PPC the object code will not change
significantly as a result of these source changes.

1) gasneti_* changes (core):
 + Rename gasneti_local_membar_force() to gasneti_local_memflush()
 + Rename gasneti_local_membar() to gasneti_local_wmb()
    On Alpha and PPC the wmb() got less costly as well.
 + Add gasneti_local_rmb() (read memory barrier)
    This is empty execpt for Alpha and PPC.

2) gasneti_* changes (extended):
 + Rename gasneti_memsyc() to gasneti_sync_writes()
 + Add gasneti_sync_reads()

3) gasnett_* changes:
 + Rename gasnett_local_membar() to gasnett_local_wmb()
 + Add gasnett_local_rmb()

4) upcri_* changes:
 + Rename upcri_local_membar() to upcri_local_wmb()
 + Add upcri_local_rmb()

5) Misc
 + testtools now tests both wmb and rmb
 + In very few places I changed a local_membar to a rmb instead
   if a wmb, or inserted both.

I tried to produce a test case in testtools that would let us know
if we were missing a required rmb on a new architecture.  However,
my attempts were unsuccesful.  In fact, the UPC-based test that was
failing on flyer all week has decided not to fail for me today :-(.

To Do:
I still need to insert calls to _rmb as needed in gasnet and upcr.
Still want a check in testtools, but that will need to come later.

IMPORTANT:
I would not normally make this big a change to the CVS HEAD, but doing
so is the best way to ensure a lot of testing (the nightlys).  I am
going to be on vacation Sat-Tues.  In case something breaks with these
changes, I've CVS tagged the changes files with
  MEMBAR_step1_pre
  MEMBAR_step1_post
to allow easy rollback of these changes (cvsmail is even better).