Fixes to Paul's membar changes Paul - please proofread and ensure you concur. This stuff is very tricky and worth careful inspection by several developers.
Improve the comments documenting the required behavior of rmb and wmb to be more explicit

Move the rmb in upcr_poll to after the gasnet_AMPoll, as it seems to make more sense there (but either is probably correct).

Fix one mpi-conduit typo (compile error)

Hoist the rmb out of the smp-conduit BLOCKUNTIL loop, so that it runs exactly once per BLOCKUNTIL, as it should eventually be for all conduits (once that's done).