fix bug 611
GASNet configure now accepts --enable-pthreads and --disable-pthreads, and the default is to probe (using a more rigorous test than previously) for working pthread support. If the pthreads test fails it's a warning by default or an error if --enable-pthreads was passed. The pthreads test requires more infrastructure than we have available at upcr configure time, so GASNet performs the test and the conduit-setup mechanism hands off the result along with the conduit list.

Also, when pthread support is missing, we disable the conduits that require it (lapi and vapi) and suppress all generation of -par and -parsync libraries and fragments.
I've verified all combinations work on lemieux's broken gcc/pthreads setup.