Fix GASNet-wide assert race in debug mode
The following assertion, OPSTATE(op) == state, that appears in the SET_OPSTATE
function presents a race when the state transition is a COMPLETION.  Other
threads spinning on a completion of the op can, once the op is seen as
complete, free it and set it in flight.

This fix also addresses the reported gm-conduit instability issues in debug
mode, which come from this assertion failure.