Fix Altix hang reported by Parry
The symptoms was upcr was hanging for all programs on startup, only with smp-conduit with 2+ pthreads on Intel C (not gcc) with -O3.

The problem was that because smp-conduit's AMPoll is a no-op, IntelC was optimizing a GASNET_BLOCKUNTIL() invocation by hoisting the condition out of the loop, resulting in a piece of code that checked the condition once then entered an infinite loop. This didn't happen for gcc because presumably it wasn't smart enough to optimize a loop containing a conditional call to the external function sched_yield().

Fix by adding a compiler optimization fence into GASNET_WAITHOOK - to ensure that very smart optimizers will never hoist the condition check out of a loop containing GASNET_WAITHOOK (which appears in all our runtime/GASNet spin-loops, via one or more macro expansions).