Add 'configure' support for disabling broken GM 2.x builds
The code and gm-conduit/README implements and describes how to deal with broken
GM 2.x builds.  Unless the user passes --enable-broken-gm, the configure script
will die if it detects 2.0 versions prior to 2.0.12 and 2.1 versions prior to
2.1.2.  Enabling the broken gm flag causes gets to be implemented as AM+puts
(as is the case with GM 1.x).

At runtime, library builds against 2.x that have gets disabled print a warning
unless the user sets GASNET_GM_NO_RDMAGET_WARNING in the environment.