Remove the serial make dependency for libgasnet All of upcr and gasnet can now safely be built using gmake -j, and the make parallelism is limited only by automake (which prevents parallelism in subdirectory recursions) and available system resources.
gmake -j still should *not* be the recommended practice for external users: because on older systems parallel make can quickly exhaust system resources, and if it encounters a compile error it's more difficult to match it up with the command that caused the problem. However, it's a nice feature for those of use who spend a significant fraction of our time waiting for gmake - my measurements indicate that adding -j to the top-level upcr gmake provides about a 2x improvement in make time on n2001, and about a 4x improvement on seaborg.

Also make a few cosmetic improvements to the libgasnet compile command which make it more amenable for cut-and-paste to the command-line for manual changes (although the recently added MANUAL_LIBCFLAGS variable should make this unnecessary in most cases).