Add some long-needed documentation on how to spawn udp-conduit jobs to the udp-conduit/README file.
Jason - please note that due to the site-specific nature of udp-conduit
spawning, the user/installer of Berkeley UPC will need to read and understand
this documentation in order to successfully use udp-conduit with Berkeley UPC
on their system (ie upcrun alone is not enough).  Our UPC docs should point
users to the gasnet/udp-conduit/README file which teaches them to configure
udp-conduit spawning for their system.