Generate a warning if tracing data is incomplete or contains duplicates.
Since the MAGIC doesn't ID the application, we don't yet detect if we have
data from multiple distinct applications.

It is still not clear what we are doing (or should be) when given data from
multiple runs of the same application.  I think as it stands now we probably do
OK if and only if the multiple runs have the same number of threads and the
same thread->node mapping.  This is reasonable, as the communications pattern
is different otherwise, making the mixing of results of very limired use.
However, we don't presently detect violations of this.