Minor modifications with regard to the use of gasnetc_max_lapi_data_size.  In the (bad) old days this value was gotten by calling the LAPI query function (LAPI_Qenv) and the value was returned through a call-by-ref arg of type int*. On 64 bit platforms, the value returned was an unsigned long and was too big to be returned as an int.
IBM (finally) depricated the use of LAPI_Qenv as a way
of returning this value and now probvides the LAPI_MAX_DATA
pre-processor symbol in the lapi.h header file.  This value
is guaranteed to fit in an unsigned long.

IBM also changed the type of the data payload length
argument to the LAPI_Put, LAPI_Get, LAPI_Amsend
from a uint to a ulong, which is consistant with
the LAPI_MAX_DATA value.