fix bug 773
AMUDP actually does support a full 256 nodes (numbered 0..255) - all 256 bit patterns of the 8-bit gasnet_node_t are legal node identifiers.

The issue here is GASNET_MAXNODES for udp-conduit is currently not representable as a gasnet_node_t. This property is not directly required by the current spec, although the lack of this property implies that gasnet_nodes() (which returns a gasnet_node_t) is not implementable when node count == GASNET_MAXNODES.

Rather than increase the size of gasnet_node_t for udp-conduit (which would affect the performance of all udp runs), I'm reducing the udp-conduit GASNET_MAXNODES to 255 to avoid this issue. udp-conduit's current buffering strategy doesn't really scale gracefully to that size job anyhow, so we're not really losing much by prohibiting 256-way udp jobs. If we ever decide large-scale udp jobs are useful enough to justify reworking AMUDP for better scalability, we can widen the udp-conduit gasnet_node_t at that time.

I'll also amend the GASNet spec to clarify this issue once we unfreeze the GASNet docs.