Refix a problem with mpicc which always wants to build w/ optimizations.
This is a problem observed on lambda (the Opteron/InfiniBand cluster at Florida)
in the past.  It was preventing us from building vapi-conduit and we worked
around it in the one file that was build with mpicc.
Well, this same problem also prevents compilation of testmpi.  So, I've switched
to a configure time probe for such a misbehaving mpicc, making the fix in the
MPI_CFLAGS if necessary.