Workaround for broken linker on Altix/ecc:
    ./configure CC=ecc

on 'ram' were running into the following error:

    ...
    checking for C compiler family... Intel
    ...
    checking for SHMEM network type... SGI/Altix
    checking for SHMEM_INCLUDE in environment... no, defaulting to
    "/usr/include"
    checking for SHMEM_LIBDIR in environment... no, defaulting to "/usr/lib"
    checking for SHMEM_LIBS in environment... no, defaulting to "-lsma"
    checking for working SHMEM configuration... no
    configure: error: build test failed: I don't know how to build SHMEM
    programs on your system

After some poking around, I determined that the problem is that ecc's linker is
broken, in that programs will not compile if '-L/usr/lib' is passed:

    $ecc hello.c                // fine
    $ecc hello.c -L/usr/lib
    /opt/intel/ecc7132/compiler70/ia64/lib/libcxa.so.4: undefined reference to `_ReadULEB'
    /opt/intel/ecc7132/compiler70/ia64/lib/libcxa.so.4: undefined reference to `_ReadSLEB'

The _ReadSLEB symbol appears to be defined in each program by the ecc linker,
and the cxa library refers to it, but then somehow the linker can't find the
symbol if /usr/lib is in the search path.

The fix for now is to set a bogus directory name when we're on the Altix:
this just gets ignored, and everything works.