Move compile-time gm-conduit parameter to be runtime-settable
GASNET_PHYSMEM_PINNABLE_RATIO can now be set in the environment to change
gm-conduit's assumption in that 70% of physical memory can be pinned in a
single job.  This hasn't caused a problem yet, but other systems may be picky
about 70% and there's no reason not to allow users to change this at runtime.
Currently, it defaults to 0.7.