Fix bug 839 - missing UPC file/line in gasnet trace
Handling the high-level line numbers correctly can be rather complicated (trace calls from header inline functio
ns being one major source of grief), so sometimes a tweak or two is required.

I've just committed a fix for the problem, and a fix to a different but related issue, that caused all pthread/d
ebug backend jobs to crash at or near startup if tracing was enabled.

This is the same patch I just applied to the release branch. The fix for
the second problem is a bit of a band-aid solution - I'll soon be adding
a more robust mechanism that will more reliably detect future such problems,
but that can wait for 2.2.