Add a new GASNet environment variable GASNET_TRACELOCAL
 GASNET_TRACELOCAL - set to control whether the PG trace messages for
   local (ie loopback) put/get operations are entered into the trace file
   or suppressed (they are included by default). This can be used to reduce
   tracing overhead and trace file size for clients with a large number of
   local put/get operations which are not of interest.

This will be used by default in Titanium to suppress the tidal wave of local put/gets that are making Titanium tracefiles so huge and consequently making gasnet_trace run so slowly on Titanium trace files. This will hopefully make it more realistic to run gasnet_trace on traces from real Titanium applications.

Teach upcr to respect the new GASNET_TRACELOCAL variable for it's tracing of local put/gets (which it must do for itself due to the GASNet bypass for loopback). Add new function upcri_getenv_yesno_withdefault to help implement this.