Add elan4 support to elan-conduit This commit adds all the configure and header goop necessary for using libelan on an elan4 system
Mostly this consisted of disabling elan3-specific features in elan-conduit when using elan4.
elan-conduit now builds and works "out-of-the-box" on thunder, an elan4/Itanium system -
all GASNet tests other than testexit pass, and simple UPC programs work.

Further work remains in tuning for elan4, and exploiting elan4-specific features.
Currently we're seeing performance like:
     put/get roundtrip latencies ~ 12us up to 2KB
          (about twice what we get on elan3, not sure why yet -
           for reference, mpi-conduit RT latency is about 28us on the same hardware)
     peak B/W of 880 MB/sec saturating around 32KB (nice)
          (mpi-conduit put/get peaks at 770/660 MB/sec around 32KB)
     anon barrier latency of 12.7us, flat up to 8 nodes (as far as I went)
          (also about twice what we get on elan3 - still investigating)
     AM ping-pong latency around 36us
Inter-thread contention in pthreaded configurations seems to be especially bad (much worse than elan3/Alpha), presumably due to the faster Itanium-2 CPU's and bus. We'll definitely need to finish the polling contention thottling work if we intend to use pthreads on elan4.
There are also some remaining issues with non-collective exits under the latest version of RMS.

So there's definitely more work to be done, but this is a reasonable first cut and should allow us to claim at least initial elan4 support in the 2.2 release.