fix bug895 - xlc crash on OSX Turns out it has nothing to do with UDP conduit - it happened on all conduits.
The culprit was an (accidental but legal) implicit integer narrowing conversion on a function call in gasnet_mmap.c, that for some bizarre reason was crashing xlc with -O3. It's definitely an optimizer bug in xlc/darwin, although the code involved was not what we really wanted anyhow.. once an explicit cast was added the problem went away - so I'm satisfied to consider this fixed, at least as far as we're concerned.
CVSMAIL: CC: