Fix a bad register usage across all the Darwin/xlC assembly code The PowerPC register convention :
http://publib.boulder.ibm.com/infocenter/pseries/index.jsp?topic=/com.ibm.aix.doc/aixassem/alangref/linkage_convent.htm

specifies that R2 is a non-volatile (callee-save) register, therefore it's inappropriate to clobber it without spilling & restoring it.
Fix our assembly code to use the volatile registers R0 and/or R4 as temporaries instead.
This was not actually causing any observable problem, but it was violating
the documented register convention and documented register restrictions
for reg_killed_by.
CVSMAIL: CC: