Add an experimental ssh-based bootstrapper to vapi-conduit. Pass --enable-vapi-bootstrap-ssh to configure to enable. With this enabled, MPI is no longer required to build vapi-conduit.
Having gotten rather upset at the reliability of mpich as a spawner
for vapi-conduit, I've started some work to implement a conduit-neutral
ssh-based spawner for gasnet.  Similiar to udp-conduit, this spawner
is "internal" to the executable.  Unlike udp-conduit, this spawner is
in C (rather than C++) and is intended to be a little more scalable
(using a "tree" of ssh commands, which is utilized for the bootstrap
barrier, broadcast and exchange, plus an additional all-to-all).

Configuration is via environment variables GASNET_SSH_CMD and
GASNET_SSH_OPTIONS, while hosts are taken from GASNET_SSH_NODEFILE
or GASNET_SSH_SERVERS (in that order).

Though currently the file is located in the vapi-conduit directory, I
am considering moving this to gasnet/other.  It contains no conduit
specific code or assumptions.  Because the "master" is a C program
rather than a script, this may also prove to be a good starting point
for TotalView support for vapi-conduit and gm-conduit.

Much more work is possible, especially porting more of the useful
features/behaviors from the udp-conduit spawner.