Initial port of GASNet and UPC to the Cray MTA
The supported conduits are smp and udp (both with pthreads).
The system has an MPI implementation, but I cannot figure out how
to make it work properly (global variables are incorrectly shared
between tasks), so for now mpi-conduit is not supported.

This shared-memory system is notable for it's unusual memory hierachy,
which is totally flat and uses no data caches - latency tolerance
is achieved by hardware context switching between threads every cycle.
This could make it an interesting target for UPC if Cray follows
through with plans to modernize the architecture with faster chips.
It's also notable for being an ILP64 system - one of the few
such architectures that appears to be in current usage (the Cray T3E
being another).

Create a "GASNet porting guide" in gasnet/README to ease future ports