Strengthen the check for broken asm/atomic.h on Linux (this does NOT fix bug 1001).
On some (all?) architectures that define it, cmpxchg is a macro with a
switch(sizeof(*ptr)), with an call to an intentionally undefined function as the
'default' case.  This is expected to compile away when building the linux kernel,
but gasnet tests may fail to link when optimizations are disabled and the linker
can't find the intentionally undefined function.
CVSMAIL: CC: