Fix bug 1000 - Add rmb barriers for Itanium After some in-depth review of the Itanium architecture manual, it appears that IA-64 does not guarantee non-conflicting reads will complete in program order, as they do in x86. Using the augmented testtools, I've seen violations of the rmb/wmb test on three different compilers for Itanium (in both debug and ndebug mode), so I'm convinced this is a feature of the architecture (as opposed to some isolated compiler bug). Adding the "mf" instruction as a rmb fixes the violations in all cases.
I suspect we may actually be able to do slightly better than the full "mf"
instruction (which is a heavyweight, bidirectional memory fence) by using
ld.acq and st.rel to dummy locations instead, as lighter-weight unidirectional
memory fence implementations of rmb and wmb. I'm continuing to
investigate this, however for the moment "mf" in each appears to be safe and
appears to effectively solve our bizarre memory behavior on Itanium.
CVSMAIL: CC: