Port GASNet to Jaquard
I have not tested UPC, because Jaquard does not have libtool installed
and it's not worth the trouble to install it since we need to fix that
silly, unnecessary libtool dependency anyhow.

All the GASNet tests pass using gcc (no changes required)
Add support for PathScale C/C++ compilers, which are installed on Jacquard.
This required tweaks to the inline assembly - PathScale does a reasonable job of handling GNU-style inline assembly, but gets it subtlely wrong. Luckily this is the only CPU they support, so we don't have to massage all our inline asm.

All the GASNet tests now pass using PathScale, except for testsmall with -O2, which mysteriously crashes in a normal printf on all conduits (no idea what that's about).

Upgrade support for PortlandGroup compilers, which are also installed on Jaquard.
PGI issues a number of bogus warnings for correct code in gasnet_extended_refcoll.c - I've already reported the bug and they claim it's fixed in the next version (6.0).
vapi-conduit is currently not supported by PGI, due to the lack of true atomics that vapi-conduit exit requires.
The recent version of PGI seems about as buggy/flaky as the last one we tried several years ago, so I'm still not overly worried about whether it works or not - I recommend avoiding it.

Fix an old issue with generic atomics and tools-only clients.
Fix a number of harmless warnings (mostly from PGI) and one honest (but harmless) type error in ref-VIS, that for some reason has gone undetected until now.
Other misc tweaks & fixes.
CVSMAIL: CC: