Initial (correct but non-optimal) fix to bug 1026
I've repaired the logic error by ensuring that the RDMA of the payload is
completed (remotely) prior to the sending of the AM header.  Note that this
only affects SEGMENT_LARGE and SEGMENT_EVERYTHING (because they use firehose)
and only for Long Request (not Replies, which are sent as Mediums).

I was able to reliably repoduce the testgasnet failure that Dan reported.
With this fix applied I am no longer able to do so.

This is a bigger hammer than really needed.  There is no need to wait for
*completion* of the RDMA.  All that is required for correctness is to
ensure that the RDMA is fully *queued* before the AM header is queued.  The
performance numbers from testam show that this fix has measureable
increased the latency for Longs.  I am working to expose the "fully queued"
event so that a more optimal solution can be implemented.
CVSMAIL: CC: