Final (correct and optimal) fix to bug 1026.
This fix is better than the previous ones in two respects.  Most importantly,
this fix doesn't have the potential for deadlock which existed in the previous
fix.  Additionally, this fix only requires that we spin wait for the RDMA
to be queued, not completed, which is the minimum for correctness.  Unlike the
initial fix, this means no spin wait is needed at all for a firehose hit.  For
a firehose miss, a shorter wait is needed than before.  AM Long throughput as
reported by testam is back to the levels measured before this bug was fixed.
CVSMAIL: CC: